refactored around log of server.

+++ local/js/eraseOrder.js
@@ -7,6 +7,12 @@ var log_warn = log4js.getLogger('warning');
 var log_crit = log4js.getLogger('critical');
 
 module.exports = function(req, res) {
+    if (req.session.user == null) {
+        res.json({ status: 255 });
+        log_warn.warn('[changeOrderState] invalid session.');
+        return;
+    }
+
     util.query(function(db) {
         db.collection('orders').deleteOne(
             { order_code: req.body.order_code },
@@ -20,7 +26,7 @@ module.exports = function(req, res) {
 
                     msg = '[eraseOrder] ' +
                           'erased order: "' + req.body.order_code + '" ' +
-                          'by "' + req.session.user_id + '".';
+                          'by "' + req.session.user.account + '".';
 
                     log_info.info(msg);
                 } else {
@@ -30,7 +36,7 @@ module.exports = function(req, res) {
                     msg = '[eraseOrder] ' +
                           'failed to erase order: "' +
                           req.body.order_code + '" ' +
-                          'by "' + req.session.user_id + '".';
+                          'by "' + req.session.user.account + '".';
 
                     log_warn.warn(msg);
                 }
